requirements.txt
torch==2.5.1
torchvision==0.20.1
pydicom==2.4.4
SimpleITK==2.3.1
Pillow==10.4.0
scikit-image==0.24.0
numpy==1.26.4
pandas==2.2.2
lxml==5.3.0
matplotlib==3.9.0
tqdm==4.67.1
jupyter==1.0.0
ipykernel==6.29.5
opencv-python==4.10.0.84

# Windows + NVIDIA GPU (optional):
# If you need CUDA 12.8 support, use the official installation command for the corresponding version:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128